+++ src/components/scores/ScoreTable.tsx
@@ -1,12 +1,13 @@
 import { getLevelText } from "../../levels/levels";
-import { GetScoresResponse } from "../../services/interfaces";
+import { GetHighScoresResponse } from "../../services/interfaces";
 
-type Props = { scores: GetScoresResponse };
+type Props = { highScores: GetHighScoresResponse };
 
-const ScoreTable = ({ scores }: Props) => (
+const ScoreTable = ({ highScores }: Props) => (
   <table className="border-separate border-spacing-1">
     <thead className="text-sm">
       <tr>
+        <th>Avatar</th>
         <th>Player</th>
         <th>Level</th>
         <th>Points</th>
@@ -14,14 +15,23 @@ const ScoreTable = ({ scores }: Props) => (
       </tr>
     </thead>
     <tbody className="text-xs">
-      {scores.highScores.map((highScore, index) => (
-        <tr key={index}>
-          <td className="text-center">{highScore.players.name}</td>
-          <td className="text-right">{getLevelText(highScore.level)}</td>
-          <td className="text-right">{highScore.points}</td>
-          <td className="text-center">{highScore.outcomes.desc}</td>
-        </tr>
-      ))}
+      {highScores.map(
+        ({ players: player, outcomes: outcome, points, level }, index) => (
+          <tr key={index}>
+            <td className="text-center">
+              <img
+                src={player.profile_pic_url}
+                alt={player.name}
+                className="w-8 h-8 rounded-full"
+              />
+            </td>
+            <td className="text-center">{player.name}</td>
+            <td className="text-right">{getLevelText(level)}</td>
+            <td className="text-right">{points}</td>
+            <td className="text-center">{outcome.desc}</td>
+          </tr>
+        )
+      )}
     </tbody>
   </table>
 );